Fix create workspace without token
